Move the test module to the logic_layer folder.
